rounding hull up to nearest int to hopefully prevent players having zero hull but being alive
git-svn-id: svn://fly.thruhere.net/forsaken@2438 1361c5d3-c82f-4d94-b015-8aedb5f389d9

+++ ProjectX/trunk/src/oct2.c
@@ -1677,7 +1677,7 @@ void DrawSimplePanel()
 			Print4x5Text( "Hull", left, top, HUDColour );	
 			if( WatchPlayerSelect.value == WhoIAm )
 			{
-				Printu_int16_t( Ships[WhoIAm].Object.Hull , right, top, HUDColour );
+				Printu_int16_t( (u_int16_t) Ships[WhoIAm].Object.Hull, right, top, HUDColour );
 				if( HullHit ) HullHit -=1;          
 			}
 			else
